Move save design and run implementation to top toolbar and remove from graph menu bar.
