+++ workbench/src/Components/ImplementationToolbar/ImplementationToolbar.js
@@ -1,9 +1,14 @@
 /* eslint-disable max-len */
-import React, {useEffect, useState} from "react";
+import React, {useCallback} from "react";
 
 import PropTypes from "prop-types";
 import {Floppy, Play} from "react-bootstrap-icons";
 import {useDispatch} from "react-redux";
+
+import {useDalEngine} from "../../Providers/GlobalProviders";
+import {useServer} from "../../Providers/GlobalProviders";
+import {setStatusMsg} from "../../Store/appSlice";
+import {useHasEntryPoint} from "../../Store/useAppSelection";
 
 import "./ImplementationToolbar.scss";
 
@@ -16,11 +21,33 @@ ImplementationToolbar.propTypes = {
  */
 export function ImplementationToolbar () {
     const dispatch = useDispatch();
+    const {engine} = useDalEngine();
+    const {sendMessage} = useServer();
+    const hasEntryPoint = useHasEntryPoint();
 
-    /**
-     * I decided to keep the toolbars separate instead of merging them into
-     * one component for now.
-     */
+    const saveDesign = useCallback(() => {
+        if (engine) {
+            engine.save();
+            dispatch(setStatusMsg("Saving design..."));
+        }
+    }, [engine, dispatch]);
+
+    const runDesign = useCallback(() => {
+        if (sendMessage && engine) {
+            if (hasEntryPoint) {
+                sendMessage({
+                    type: "terminal_run_entry_point",
+                    data: engine.implementation.getEntryPoint(),
+                });
+            } else {
+                const failureMsg = "Failed to run design. Please ensure an entry point is set.";
+                sendMessage({
+                    type: "terminal_run_entry_point",
+                    data: `echo "${failureMsg}"`,
+                });
+            }
+        }
+    }, [sendMessage, hasEntryPoint, engine]);
 
     return (
         <div className="mainToolBar">
@@ -29,14 +56,14 @@ export function ImplementationToolbar () {
                 <div className="mainToolBarGroup"></div>
             </div>
             <div className="mainToolBarRight">
-                <span className="mainToolBarButton" >
+                <span className="mainToolBarButton" onClick={runDesign}>
                     <Play
                         size={20}
                         style={{"color": "white", "cursor": "pointer"}}/>
                     <span className="mainToolBarButton">Run Implementation</span>
                 </span>
 
-                <span className="mainToolBarButton" >
+                <span className="mainToolBarButton" onClick={saveDesign}>
                     <Floppy
                         size={14}
                         style={{"color": "white", "cursor": "pointer"}}/>
